fix: add test suites, audit READMEs, fix API contract inconsistencies
- Add comprehensive test suites for all 5 backends (149 tests, 100% pass)
  - Spring Boot: PostgreSQL profile (22 tests)
  - Node.js: Jest + supertest (37 tests)
  - FastAPI: pytest + httpx.AsyncClient (19 tests)
  - Laravel: PHPUnit + DatabaseTransactions (34 tests)
  - Plain PHP: curl + php -S integration (37 tests)
- Audit and fix all 7 READMEs against actual source code (26 fixes)
- Fix FastAPI Decimal serialization: balance/amount now float not string
- Fix Plain PHP: page default 0→1, PATCH→PUT, /stats→/transactions/stats
- Fix WalletResponse/TransactionResponse schemas for frontend compatibility
- Spring Boot tests: H2→PostgreSQL (pg-test profile)

+++ digital_wallet_fastapi/app/schemas/admin.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from decimal import Decimal
 from typing import TypeVar, Generic
 
 from pydantic import BaseModel, Field, ConfigDict
@@ -30,7 +29,7 @@ class AdminTransactionResponse(BaseModel):
     id: int
     from_wallet_id: int | None = Field(alias="fromWalletId", default=None)
     to_wallet_id: int | None = Field(alias="toWalletId", default=None)
-    amount: Decimal
+    amount: float
     tx_type: str = Field(alias="txType")
     status: str
     created_at: datetime = Field(alias="createdAt")
@@ -42,11 +41,11 @@ class DailyVolumeItem(BaseModel):
     model_config = ConfigDict(populate_by_name=True)
     date: str
     count: int
-    amount: Decimal
+    amount: float
 
 
 class TransactionStatsResponse(BaseModel):
     model_config = ConfigDict(populate_by_name=True)
     total_transactions: int = Field(alias="totalTransactions")
-    total_amount: Decimal = Field(alias="totalAmount")
+    total_amount: float = Field(alias="totalAmount")
     daily_volume: list[DailyVolumeItem] = Field(alias="dailyVolume")
